refactor(Board)[OK-89]: Split fabricCanvasUtils into multiple context-specific files. Moved interfaces, reducers, and types to dedicated Board-related folders. Added comments for clarity.
