+++ frontend/src/enums/CanvasObjectTypes.ts
@@ -1,0 +1,9 @@
+export enum CanvasObjectTypes {
+  Rect = "rect",
+  Circle = "circle",
+  Line = "line",
+  Path = "path",
+  Image = "image",
+  IText = "i-text",
+  ActiveSelection = "activeSelection",
+}
