+++ frontend/src/enums/CanvasReducerAction.ts
@@ -1,0 +1,8 @@
+export enum CanvasReducerActionEnum {
+  SET_CANVAS = "SET_CANVAS",
+  SET_CANVAS_MODE = "SET_CANVAS_MODE",
+  SET_COLOR = "SET_COLOR",
+  SET_SIZE = "SET_SIZE",
+  SET_SELECTED_OBJECT_STYLES = "SET_SELECTED_OBJECT_STYLES",
+  SET_ACTIVE_ITEM = "SET_ACTIVE_ITEM",
+}
