+++ frontend/src/interfaces/canvas-context.ts
@@ -1,0 +1,17 @@
+import { CanvasReducerState } from "@/types/CanvasReducer";
+import { MutableRefObject } from "react";
+import { CanvasMode } from "@/enums/CanvasMode";
+
+export interface ICanvasContext {
+  state: CanvasReducerState;
+  canvasRef: MutableRefObject<HTMLCanvasElement | null>;
+  setCanvasMode: (mode: CanvasMode) => void;
+  setColor: (color: string) => void;
+  setSize: (size: number) => void;
+  handleStyleChange: (styles: object) => void;
+  setActiveItem: (activeItem: string | null) => void;
+}
+
+export interface CanvasRef {
+  current: HTMLCanvasElement | null;
+}
