+++ frontend/src/interfaces/undo-redo-context.ts
@@ -1,0 +1,5 @@
+export interface IUndoRedoContext {
+  saveState: () => void;
+  undo: () => void;
+  redo: () => void;
+}
