--- frontend/src/utils/isActiveItem.ts
@@ -1,5 +1,0 @@
-import { CanvasMode } from "@/enums/CanvasMode";
-
-export const isActiveItem = (itemName: string, activeItem: string | null, canvasMode: CanvasMode): boolean => {
-  return itemName === activeItem || itemName === canvasMode.toString();
-};
